+++ config/splat.dra.yaml
@@ -73,6 +73,5 @@ segments:
       - [0x75DA0, c, 75DA0]
       - [0x75F54, c, 75F54]
       - [0x953A0, c, 953A0]
-      - [0x95484, c, 95484]
       - [0x962AC, data]
   - [0x119870]
