+++ src/dra/953A0.c
@@ -39,3 +39,59 @@ void func_801353A0(void) {
         break;
     }
 }
+
+INCLUDE_ASM("asm/dra/nonmatchings/95484", func_80135484);
+
+void func_80135624(s16 arg0, s32 arg1, s32 arg2, s16 arg3, s16 arg4) {
+    if (arg2 != 0) {
+        D_80138F28 |= (1 << ((arg1 + 6) * 2)) + (1 << (((arg1 + 6) * 2) + 1));
+    }
+    func_80132A04((arg1 * 2) + 12, D_800BF554[arg0 * 7], D_800BF555[arg0 * 7],
+                  D_800BF559[arg0 * 7], D_800BF556[arg0 * 7], arg3, arg4);
+    do { // TODO: !FAKE
+    } while (0);
+    D_8013B650[arg1] = arg0;
+    D_8013AED4[arg1] = D_800BF55A[arg0 * 7];
+}
+
+INCLUDE_ASM("asm/dra/nonmatchings/95484", func_8013572C);
+
+s16 func_80135C00(s16 arg0) {
+    arg0++;
+
+    if (arg0 == 0x100) {
+        arg0 = 0;
+    }
+
+    return arg0;
+}
+
+INCLUDE_ASM("asm/dra/nonmatchings/95484", func_80135C2C);
+void func_80135C2C();
+
+INCLUDE_ASM("asm/dra/nonmatchings/95484", func_80135D8C);
+void func_80135D8C();
+
+INCLUDE_ASM("asm/dra/nonmatchings/95484", func_80136010);
+
+void func_801361F8(void) {
+    if (D_8013AEEC != 0) {
+        func_80136010();
+        func_80135C2C();
+        func_80135D8C();
+        func_80134F50();
+        func_80133FCC();
+        func_801353A0();
+        func_8002A024(0, D_80138F28);
+        D_80138F28 = 0;
+        func_80131FCC();
+
+        if (D_80139020) {
+            func_801324B4(0, 0, 0);
+        } else {
+            func_801324B4(0, D_8013B668, D_8013B668);
+        }
+    }
+}
+
+void nullsub_10(void) {}
